refactor: edit constants style
