+++ app.js
@@ -2,7 +2,6 @@ require("dotenv").config();
 
 const express = require("express");
 const path = require("path");
-const cookieParser = require("cookie-parser");
 const logger = require("morgan");
 
 require("./config/db");
@@ -21,13 +20,11 @@ nodeSchedule();
 app.use(logger("dev"));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
-app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
 
 app.get("/", (req, res) => {
   res.json({ message: "success" });
 });
-
 app.use("/users", usersRouter);
 app.use("/pins", pinsRouter);
