+++ middlewares/handleError.js
@@ -1,11 +1,11 @@
-const ERROR = require("../constants/error");
+const { ERROR } = require("../constants");
 
 function handleError(err, req, res, next) {
   res.status(err.status || 500);
 
   res.json({
     code: err.status,
-    message: err.message ? err.message : ERROR.SERVER.default,
+    message: err.message ? err.message : ERROR.server,
   });
 }
