+++ package.json
@@ -10,8 +10,6 @@
     "aws-sdk": "^2.983.0",
     "buffer-from": "^1.1.2",
     "busboy": "^0.3.1",
-    "cookie-parser": "^1.4.5",
-    "cors": "^2.8.5",
     "dayjs": "^1.10.7",
     "debug": "~2.6.9",
     "dotenv": "^10.0.0",
