+++ utils/getDate.js
@@ -2,13 +2,15 @@ const dayjs = require("dayjs");
 const duration = require("dayjs/plugin/duration");
 dayjs.extend(duration);
 
+const { PIN_COUNT } = require("../constants");
+
 function getDate(time) {
   const targetTime = dayjs(time);
   const currentTime = dayjs(new Date());
 
   const difference = currentTime.diff(targetTime, "second");
 
-  if (difference >= 86400) {
+  if (difference >= PIN_COUNT) {
     return true;
   }
